+++ KeepGardeningProject_customer/app/src/main/java/com/test/keepgardeningproject_customer/UI/ProductCustomerDetail/ProductCustomerDetailBottomDialog.kt
@@ -1,4 +1,4 @@
-package com.test.keepgardeningproject_customer.UI.ProductCustomerDetail.components
+package com.test.keepgardeningproject_customer.UI.ProductCustomerDetail
 
 import android.os.Bundle
 import android.view.LayoutInflater
