+++ KeepGardeningProject_customer/app/src/main/res/layout/fragment_product_customer_detail.xml
@@ -72,7 +72,7 @@
                 android:textStyle="bold" />
 
             <TextView
-                android:id="@+id/textView"
+                android:id="@+id/textView_pcd_price"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
                 android:layout_marginTop="15dp"
@@ -83,18 +83,28 @@
                 android:textSize="18sp"
                 android:textStyle="bold" />
 
-            <com.google.android.material.tabs.TabLayout
-                android:id="@+id/tabs_pcd"
+            <View
+                android:id="@+id/divider"
+                android:layout_width="match_parent"
+                android:layout_height="4dp"
+                android:layout_marginTop="15dp"
+                android:background="?android:attr/listDivider" />
+
+            <TextView
+                android:id="@+id/textView_pcd_info"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:layout_marginTop="15dp">
+                android:layout_margin="15dp"
+                android:text="상품상세 설명입니다.\n줄바꿈내용"
+                android:textColor="@color/black"
+                android:textSize="16sp" />
 
-            </com.google.android.material.tabs.TabLayout>
-
-            <androidx.viewpager2.widget.ViewPager2
-                android:id="@+id/viewPager_pcd_tab"
+            <androidx.recyclerview.widget.RecyclerView
+                android:id="@+id/recyclerView_pcd"
                 android:layout_width="match_parent"
-                android:layout_height="wrap_content" />
+                android:layout_height="match_parent"
+                android:layout_marginTop="25dp" />
+
         </LinearLayout>
     </androidx.core.widget.NestedScrollView>
